Refactor required fields
