+++ package.json
@@ -7,9 +7,9 @@
     "start-ios": "react-native run-ios",
     "start-android": "react-native run-android",
     "convert-androidx": "npx jetify",
-    "test": "jest",
+    "test": "jest --maxWorkers=4",
     "lint-staged": "lint-staged",
-    "test:coverage": "jest --coverage"
+    "test:coverage": "jest --maxWorkers=4 --coverage"
   },
   "dependencies": {
     "@delightfulstudio/react-native-wheel-picker-android": "^2.1.3",
@@ -77,6 +77,7 @@
     "redux-mock-store": "^1.5.3"
   },
   "jest": {
+    "testEnvironment": "node",
     "preset": "react-native",
     "setupFilesAfterEnv": [
       "<rootDir>/testing/config.js"
